# applications/rpg_table_helper/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Main API v1.0",
    "version": "v1.0"
  },
  "paths": {
...
    "/SignIn/testlogin": {
      "get": {
        "tags": [
          "SignIn"
        ],
        "summary": "This method returns \"Welcome\" when the provided JWT is valid.",
        "responses": {
          "200": {
            "description": "Returns the JWT for the user.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "If the used jwt is not valid",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/SignIn/getloginchallengeforusername/{username}": {
      "post": {
        "tags": [
          "SignIn"
        ],
        "summary": "Returns the encryption challenge for a given username.",
        "description": "Please note that the client has to provide an RSA public\r\nkey in order to receive the encrypted challenge dictionary.",
        "parameters": [{
          "name": "username",
          "in": "path",
          "description": "The username of the desired encryption challenge",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "requestBody": {
          "description": "The public RSA key of the client",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/EncryptedMessageWrapperDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncryptedMessageWrapperDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EncryptedMessageWrapperDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EncryptedMessageWrapperDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns the encrypted challenge dict for the user.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "If request is not valid (user not found, key not valid, etc.)",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/SignIn/login": {
      "post": {
        "tags": [
          "SignIn"
        ],
        "summary": "Performs the login with username and password.",
        "requestBody": {
          "description": "The username and password",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginWithUsernameAndPasswordDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginWithUsernameAndPasswordDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginWithUsernameAndPasswordDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginWithUsernameAndPasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns the jwt for the user",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "If username or password did not match",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/SignIn/loginwithapple": {
      "post": {
        "tags": [
          "SignIn"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/AppleLoginDetails"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleLoginDetails"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleLoginDetails"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AppleLoginDetails"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/SignIn/loginwithgoogle": {
      "post": {
        "tags": [
          "SignIn"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/SignIn/requestresetpassword": {
      "post": {
        "tags": [
          "SignIn"
        ],
        "summary": "Requests an reset password email for the user.",
        "requestBody": {
          "description": "The username and email for the user requesting the reset",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequestDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns the text \"Email sent\"",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "If username or password did not match",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/SignIn/resetpassword": {
      "post": {
        "tags": [
          "SignIn"
        ],
        "summary": "Completes the password reset requests by providing the reset code from the email.",
        "requestBody": {
          "description": "The username and email for the user requesting the reset",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns the text \"New password set\"",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "If username, email or reset code did not match",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AppleLoginDetails": {
        "required": [
          "authorizationCode",
          "identityToken"
        ],
        "type": "object",
        "properties": {
          "authorizationCode": {
            "minLength": 1,
            "type": "string"
          },
          "identityToken": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "EncryptedMessageWrapperDto": {
        "type": "object",
        "properties": {
          "encryptedMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Represents a data transfer object for encapsulating encrypted messages."
      },
...
      "GoogleLoginDto": {
        "required": [
          "accessToken",
          "identityToken"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "minLength": 1,
            "type": "string",
            "description": "Gets or sets the Google Access Token."
          },
          "identityToken": {
            "minLength": 1,
            "type": "string",
            "description": "Gets or sets the Google Identity Token."
          }
        },
        "additionalProperties": false,
        "description": "Data Transfer Object for Google login information."
      },
...
      "HttpValidationProblemDetails": {
        "allOf": [{
            "$ref": "#/components/schemas/ProblemDetails"
          },
          {
            "type": "object",
            "properties": {
              "errors": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "nullable": true
              }
            },
            "additionalProperties": {}
          }
        ]
      },
...
      "LoginWithUsernameAndPasswordDto": {
        "required": [
          "username",
          "userSecretByEncryptionChallenge"
        ],
        "type": "object",
        "properties": {
          "username": {
            "minLength": 1,
            "type": "string",
            "description": "Gets or sets the username of the user."
          },
          "userSecretByEncryptionChallenge": {
            "minLength": 1,
            "type": "string",
            "description": "Gets or sets the hashed password of the user.\r\nEither this has to be set or the ApiKey.\r\n            \r\nIf this is set, the user is registering in with username + password, if not set,\r\nthe user is providing username and password after the sign in using a OIDC Provider\r\n(like apple or google)."
          }
        },
        "additionalProperties": false,
        "description": "Data Transfer Object for user login information."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "ResetPasswordRequestDto": {
        "required": [
          "email",
          "username"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "description": "Gets or sets the email address of the user requesting a password reset.",
            "format": "email"
          },
          "username": {
            "minLength": 1,
            "type": "string",
            "description": "Gets or sets the username of the user requesting a password reset."
          }
        },
        "additionalProperties": false,
        "description": "Data Transfer Object for password reset requests."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 1safsfsdfdfd\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [{
    "Bearer": []
  }]
}